feat: 해설 목록을 위한 전역 상태 추가 및 localStorageEffect 추가
